requirements-dev.txt
black==23.3.0
pytest==7.3.2
python-dotenv==1.0.0
ruff==0.0.272